fix for x86-64
Originally committed as revision 8022 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/i386/dsputil_h264_template_mmx.c
@@ -246,7 +246,7 @@ static void H264_CHROMA_MC4_TMPL(uint8_t *dst/*align 4*/, uint8_t *src/*align 1*
         "sub $2, %2                 \n\t"
         "jnz 1b                     \n\t"
         : "+r"(dst), "+r"(src), "+r"(h)
-        : "r"(stride), "m"(ff_pw_32), "m"(x), "m"(y), "m"(ff_pw_8)
+        : "r"((long)stride), "m"(ff_pw_32), "m"(x), "m"(y), "m"(ff_pw_8)
     );
 }
